split artifacts, cleanup
