+++ src/compiler/r1cs_gen.rs
@@ -1,4 +1,4 @@
-use std::{cell::RefCell, fmt::Display, rc::Rc};
+use std::{cell::RefCell, rc::Rc};
 
 use crate::compiler::{
     analysis::{
@@ -9,39 +9,9 @@ use crate::compiler::{
     ssa::{BinaryArithOpKind, BlockId, CmpKind, FunctionId, MemOp, Radix, SSA, SliceOpDir},
 };
 use ark_ff::{AdditiveGroup, BigInt, BigInteger, Field, PrimeField};
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
-use tracing::{error, instrument, warn};
+use tracing::{instrument, warn};
 
-mod lc_serde {
-    use super::*;
-
-    pub fn serialize<S>(lc: &Vec<(usize, ark_bn254::Fr)>, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let converted: Vec<(usize, [u64; 4])> = lc
-            .iter()
-            .map(|(idx, coeff)| (*idx, coeff.into_bigint().0))
-            .collect();
-        converted.serialize(serializer)
-    }
-
-    pub fn deserialize<'de, D>(deserializer: D) -> Result<Vec<(usize, ark_bn254::Fr)>, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let converted: Vec<(usize, [u64; 4])> = Deserialize::deserialize(deserializer)?;
-        Ok(converted
-            .into_iter()
-            .map(|(idx, limbs)| {
-                (
-                    idx,
-                    ark_bn254::Fr::from_bigint(BigInt(limbs)).expect("Invalid field element"),
-                )
-            })
-            .collect())
-    }
-}
+pub use mavros_artifacts::{ConstraintsLayout, LC, R1C, R1CS, WitnessLayout};
 
 // #[derive(Clone, Debug, Copy, PartialEq, PartialOrd, Eq, Ord)]
 // pub enum WitnessIndex {
@@ -50,8 +20,6 @@ mod lc_serde {
 //     LookupValueInverse(usize),
 //     LookupValueInverseAux(usize),
 // }
-
-type LC = Vec<(usize, crate::compiler::Field)>;
 
 #[derive(Clone, Debug)]
 struct ArrayData {
@@ -278,16 +246,6 @@ impl Value {
     }
 }
 
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct R1C {
-    #[serde(with = "lc_serde")]
-    pub a: LC,
-    #[serde(with = "lc_serde")]
-    pub b: LC,
-    #[serde(with = "lc_serde")]
-    pub c: LC,
-}
-
 #[derive(Clone, Debug)]
 pub struct LookupConstraint {
     pub table_id: usize,
@@ -298,39 +256,6 @@ pub struct LookupConstraint {
 pub enum Table {
     Range(u64),
     OfElems(Vec<LC>),
-}
-
-fn field_to_string(c: ark_bn254::Fr) -> String {
-    if c.into_bigint() > crate::compiler::Field::MODULUS_MINUS_ONE_DIV_TWO {
-        format!("-{}", -c)
-    } else {
-        c.to_string()
-    }
-}
-
-impl Display for R1C {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let a_str = self
-            .a
-            .iter()
-            .map(|(i, c)| format!("{} * v{}", field_to_string(*c), i))
-            .collect::<Vec<_>>()
-            .join(" + ");
-        let b_str = self
-            .b
-            .iter()
-            .map(|(i, c)| format!("{} * v{}", field_to_string(*c), i))
-            .collect::<Vec<_>>()
-            .join(" + ");
-        let c_str = self
-            .c
-            .iter()
-            .map(|(i, c)| format!("{} * v{}", field_to_string(*c), i))
-            .collect::<Vec<_>>()
-            .join(" + ");
-
-        write!(f, "({}) * ({}) - ({}) = 0", a_str, b_str, c_str)
-    }
 }
 
 #[derive(Clone)]
@@ -741,15 +666,6 @@ impl symbolic_executor::Value<R1CGen> for Value {
     }
 }
 
-pub use crate::vm::{ConstraintsLayout, WitnessLayout};
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct R1CS {
-    pub witness_layout: WitnessLayout,
-    pub constraints_layout: ConstraintsLayout,
-    pub constraints: Vec<R1C>,
-}
-
 impl R1CGen {
     pub fn new() -> Self {
         Self {
@@ -1010,139 +926,5 @@ impl R1CGen {
             constraints_layout,
             constraints: result,
         };
-    }
-}
-
-impl R1CS {
-    pub fn compute_derivatives(
-        &self,
-        coeffs: &[crate::compiler::Field],
-        res_a: &mut [crate::compiler::Field],
-        res_b: &mut [crate::compiler::Field],
-        res_c: &mut [crate::compiler::Field],
-    ) {
-        for (r1c, coeff) in self.constraints.iter().zip(coeffs.iter()) {
-            for (a_ix, a_coeff) in r1c.a.iter() {
-                res_a[*a_ix] += *a_coeff * *coeff;
-            }
-            for (b_ix, b_coeff) in r1c.b.iter() {
-                res_b[*b_ix] += *b_coeff * *coeff;
-            }
-            for (c_ix, c_coeff) in r1c.c.iter() {
-                res_c[*c_ix] += *c_coeff * *coeff;
-            }
-        }
-    }
-
-    pub fn check_witgen_output(
-        &self,
-        pre_comm_witness: &[crate::compiler::Field],
-        post_comm_witness: &[crate::compiler::Field],
-        a: &[crate::compiler::Field],
-        b: &[crate::compiler::Field],
-        c: &[crate::compiler::Field],
-    ) -> bool {
-        let witness = [pre_comm_witness, post_comm_witness].concat();
-        if a.len() != self.constraints_layout.size() {
-            error!(message = %"The a vector has the wrong length", expected = self.constraints_layout.size(), actual = a.len());
-            return false;
-        }
-        if b.len() != self.constraints_layout.size() {
-            error!(message = %"The b vector has the wrong length", expected = self.constraints_layout.size(), actual = b.len());
-            return false;
-        }
-        if c.len() != self.constraints_layout.size() {
-            error!(message = %"The c vector has the wrong length", expected = self.constraints_layout.size(), actual = c.len());
-            return false;
-        }
-        for (i, r1c) in self.constraints.iter().enumerate() {
-            let av = r1c
-                .a
-                .iter()
-                .map(|(i, c)| c * &witness[*i])
-                .sum::<ark_bn254::Fr>();
-
-            let bv = r1c
-                .b
-                .iter()
-                .map(|(i, c)| c * &witness[*i])
-                .sum::<ark_bn254::Fr>();
-
-            let cv = r1c
-                .c
-                .iter()
-                .map(|(i, c)| c * &witness[*i])
-                .sum::<ark_bn254::Fr>();
-            let mut fail = false;
-            if av * bv != cv {
-                error!(message = %"R1CS constraint failed to verify", index = i);
-                fail = true;
-            }
-            if av != a[i] {
-                error!(message = %"Wrong A value for constraint", index = i, actual = a[i].to_string(), expected = av.to_string());
-                fail = true;
-            }
-            if bv != b[i] {
-                error!(message = %"Wrong B value for constraint", index = i, actual = b[i].to_string(), expected = bv.to_string());
-                fail = true;
-            }
-            if cv != c[i] {
-                error!(message = %"Wrong C value for constraint", index = i, actual = c[i].to_string(), expected = cv.to_string());
-                fail = true;
-            }
-            if fail {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    pub fn check_ad_output(
-        &self,
-        coeffs: &[crate::compiler::Field],
-        a: &[crate::compiler::Field],
-        b: &[crate::compiler::Field],
-        c: &[crate::compiler::Field],
-    ) -> bool {
-        let mut a = a.to_vec();
-        let mut b = b.to_vec();
-        let mut c = c.to_vec();
-        for (r1c, coeff) in self.constraints.iter().zip(coeffs.iter()) {
-            for (a_ix, a_coeff) in r1c.a.iter() {
-                a[*a_ix] -= *a_coeff * *coeff;
-            }
-            for (b_ix, b_coeff) in r1c.b.iter() {
-                b[*b_ix] -= *b_coeff * *coeff;
-            }
-            for (c_ix, c_coeff) in r1c.c.iter() {
-                c[*c_ix] -= *c_coeff * *coeff;
-            }
-        }
-        let mut wrongs = 0;
-        for i in 0..a.len() {
-            if a[i] != crate::compiler::Field::ZERO {
-                if wrongs == 0 {
-                    error!(message = %"Wrong A deriv for witness", index = i);
-                }
-                wrongs += 1;
-            }
-            if b[i] != crate::compiler::Field::ZERO {
-                if wrongs == 0 {
-                    error!(message = %"Wrong B deriv for witness", index = i);
-                }
-                wrongs += 1;
-            }
-            if c[i] != crate::compiler::Field::ZERO {
-                if wrongs == 0 {
-                    error!(message = %"Wrong C deriv for witness", index = i);
-                }
-                wrongs += 1;
-            }
-        }
-        if wrongs > 0 {
-            error!("{} out of {} wrong derivatives", wrongs, 3 * a.len());
-            return false;
-        }
-        return true;
     }
 }
