+++ src/lib.rs
@@ -1,13 +1,12 @@
 pub mod abi_helpers;
-pub mod compiled_artifacts;
 pub mod compiler;
 pub mod driver;
 pub mod error;
 pub mod project;
 
+pub use mavros_artifacts as artifacts;
 pub use mavros_vm as vm;
 
-pub use compiled_artifacts::CompiledArtifacts;
 pub use error::Error;
 pub use project::Project;
