+++ vm/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2024"
 [dependencies]
 ark-bn254 = { version = "0.5.0" }
 ark-ff = { version = "0.5.0" }
+mavros-artifacts = { path = "../mavros-artifacts" }
 noirc_abi = { git = "https://github.com/noir-lang/noir.git", branch = "master", package = "noirc_abi" }
 opcode-gen = { path = "../opcode-gen" }
 plotters = { version = "0.3.5" }
